Reorganize files to accomodate Nuxt component discovery and exclude `.types` files from it
