+++ src/components/VModal/VModalTarget.vue
@@ -4,7 +4,7 @@
 
 <script>
 import { defineComponent } from '@nuxtjs/composition-api'
-import { VTeleportTarget } from '~/components/VTeleport'
+import VTeleportTarget from '~/components/VTeleport/VTeleportTarget'
 
 export default defineComponent({
   name: 'VModalTarget',
