+++ src/components/VTeleport/README.md
@@ -1,0 +1,3 @@
+# VTeleport and VTeleportTarget
+
+Directly based on code from https://gist.github.com/Alan-Liang/a847559433919ecb53dff217ed9708bb
