+++ src/components/VTeleport/meta/targets.js
@@ -1,0 +1,4 @@
+/**
+ * @type {Record<string, import('vue').Component>}
+ */
+export const targets = {}
